--- backend/app/__main__.py
@@ -1,13 +1,0 @@
-import uvicorn
-import os
-import logging
-
-logging.basicConfig(filename='backend.log', level=logging.DEBUG)
-
-
-def main():
-    uvicorn.run("app:app", host='0.0.0.0', port=8765, reload=True, log_level='info', workers=os.cpu_count())
-
-
-if __name__ == "__main__":
-    main()
