+++ backend/app/routers/__init__.py
@@ -1,0 +1,8 @@
+from . import general
+
+
+
+routers = (
+	general.router,
+
+)
